requirements.txt
ddgs==9.9.3
python-dotenv==1.2.1
PyYAML==6.0.2
torch==2.9.1+cu126
transformers==4.57.3
